Updated
- Significantly improved performance of the miniature projection.
NOTE: The miniature projection is a separate entity and it takes time to
load, depending on the ship size.
NOTE: The ship edges are visible and fully opaque, I can't find a way to
disable those.
- Improved performance of status mode by coloring a maximum of 10 blocks
per tick and checking 50 per tick.
- Added UI config which replaces the tag name system, name is still used
for storage but should not be edited manually to change settings.
- Added legacy tag name system support so your settings aren't
completely lost.
NOTE: The offset is the only thing not read from the old system, you'll
need to reconfigure that.
- Added fixed rotation in all 3 axis as well as spinning in all 3 axis.
- Changed the rendering distance check to use camera position instead of
player position.