Re-done entirely.
- Changed how preview projection is created.
It is now a separate entity created by the mod and the projector is
simply removed of its projection when it's in preview mode.
This does not provide the best performance as it requires deleting and
spawning grids between preview/build mode changes, but there's no other
way right now.
- Spawning and saving of grids is done in another thread to not hitch
the game too much.
- Settings are now stored internally in the block, no more weird names!
The previous settings from the name are being read by the mod and
converted to the new system.
There might be issues with synchronizing the settings since it's a new
system, please be aware of such posibility and report issues that you
find with as much detail as possible.
- Re-done the terminal UI, separated each mode's controls per category
(preview or build mode).
- Toolbar actions and PB properties are now available for this mod's
controls.
This is because the terminal controls system was fixed for projectors.
- Added the ability to load the ship that the projector is on as a
blueprint copy.
- Updated status mode to cover more differences and special states are
being highlighted (seen above other transparent blocks) apart from being
colored; special states are being shown in the detailed info.
- Changed rotate sliders to be positive or negative and instead added a
checkbox for toggling spin.
- Reworked how the blueprint is rotated, by priority of pitch then yaw
then roll, this should prevent issues that caused pitch and roll to
rotate on the same axis for example.
- Added lights to the preview projection to imitate a hologram better.
- Fixed scaling to better fit the projection inside the size you specify
in the slider.
- Added see-through mode that makes armor and decorative blocks less
visible, causing other internal blocks to be seen.